Clone protoc from csharp dir, but don't use grpc stuff
grpc is not yet ready for dotnet core, although protobuf is. For whatever
reason, when generating the protobuf code with the grpc plugins included, the
generated code is slightly different in one minor way. Here is an example:
```
+        errorMessage_ = pb::Preconditions.CheckNotNull(value, "value");
-        errorMessage_ = pb::ProtoPreconditions.CheckNotNull(value, "value");
```
The `-` part of this is the generated code with the grpc mixins, the `+` part is
without it.